Don't show support button for your own project
